+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 * Add `blikAliases` to LocalPaymentCompleted webhook and LocalPaymentDetails
 * Deprecate `SamsungPayCard`
+* Updated expiring pinned vendor SSL certificates
 
 ## 3.36.0
 * Add `payerName`, `bic` and `iban_last_chars` for LocalPaymentCompleted webhook